% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{expanse_cols}
\alias{expanse_cols}
\alias{expanse_palette}
\title{Colors from The Expanse (WIP)}
\format{An object of class \code{list} of length 14.}
\usage{
expanse_cols

expanse_palette
}
\description{
Colors from The Expanse (WIP)
}
\keyword{datasets}
