% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{expanse_pal}
\alias{expanse_pal}
\title{Color palette based on selected colors from \link{expanse_cols}}
\usage{
expanse_pal()
}
\description{
Color palette based on selected colors from \link{expanse_cols}
}
