% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{protomolecule}
\alias{protomolecule}
\title{Helper to get the right font name for varous operating systems}
\usage{
protomolecule(weight = c("light", "regular", "medium", "bold", "black"))
}
\arguments{
\item{weight}{the weight of the font}
}
\description{
Due to the way the font metadata has been crafted, the "PostScript Family Name"
may be different on any given operating system. This is a helper to
get the right name.
}
