% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-expanse.R
\name{theme_expanse}
\alias{theme_expanse}
\title{An "Expanse" theme for {ggplot2}}
\usage{
theme_expanse(base_family = protomolecule("light"), base_size = 11.5,
  plot_title_family = protomolecule("medium"), plot_title_size = 18,
  plot_title_face = "plain", plot_title_margin = 10,
  plot_title_col = expanse_cols$white,
  subtitle_family = protomolecule("light"), subtitle_size = 13,
  subtitle_face = "plain", subtitle_col = expanse_cols$mcrn_red_1,
  subtitle_margin = 15, strip_text_family = base_family,
  strip_text_size = 12, strip_text_face = "plain",
  strip_text_col = expanse_cols$white,
  caption_family = protomolecule("light"), caption_size = 9,
  caption_face = "plain", caption_margin = 10,
  caption_col = expanse_cols$white, axis_text_size = base_size,
  axis_title_family = base_family, axis_title_size = 9,
  axis_title_face = "plain", axis_title_just = "rt",
  plot_margin = margin(30, 30, 30, 30), panel_spacing = grid::unit(2,
  "lines"), text_col = expanse_cols$white,
  grid_col = alpha(expanse_cols$white, 1/2), axis_col = grid_col,
  background_col = expanse_cols$background,
  foreground_col = expanse_cols$display_blue, grid = TRUE,
  axis = FALSE, ticks = FALSE)
}
\arguments{
\item{base_family, base_size}{base font family and size}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin, plot_title_col}{plot tilte family, face, size, margin, and color}

\item{subtitle_family, subtitle_face, subtitle_size, subtitle_col}{plot subtitle family, face, size, and color}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size, strip_text_col}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin, caption_col}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size}{axis title font family, face and size}

\item{axis_title_just}{axis title font justificationk one of \code{[blmcrt]}}

\item{plot_margin}{plot margin (specify with \link[ggplot2:margin]{ggplot2::margin})}

\item{panel_spacing}{panel spacing (use \code{unit()})}

\item{text_col, axis_col, background_col, foreground_col}{more colors to set}

\item{grid_col}{grid color}

\item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y})}

\item{axis}{add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}"}

\item{ticks}{ticks if \code{TRUE} add ticks}
}
\description{
You should \url{import_protomolecule} first and also install the fonts on your
system before trying to use this theme.
}
\details{
There is an option \code{hrbrthemes.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
