% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggexpanse-package.R
\docType{package}
\name{ggexpanse}
\alias{ggexpanse}
\alias{ggexpanse-package}
\title{Theme Elements Based On 'The Expanse'}
\description{
Plot aesthetic defaults and theme elements inspired by 'The Expanse'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.com/hrbrmstr/ggexpanse}
  \item Report bugs at \url{https://gitlab.com/hrbrmstr/ggexpanse/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
